feat: ngram frequency calculations (tf-idf)

+++ pkg/keywords/keyword.go
@@ -2,5 +2,5 @@ package keywords
 
 type Keyword interface {
 	Keyword() string
-	Weight() float32
+	Weight() float64
 }
